This change makes GeLU go 8x faster on Intel, 3x faster on Apple Silicon, and 2x faster on Threadripper. It's the world's most popular activation function, crucial to models such as Whisper and Gemma. On those models, this change can have a noticeable improvement in performance. That's because GeLU is usually the most time-consuming op except for matrix multiplication

In addition to improving performance, this change also improves accuracy. On ARM64 and AMD64 systems, we no longer need to rely on a 16-bit lookup table. We're now using SIMD instead. The GeLU lookup table is still here except it's been converted from fp16 to bf16. This helps align inference more with training possibly, but it helps us avoid the two extra lookups into the fp16 table. Therefore this change should have a positive impact on performance for platforms like OpenPOWER and RISC-V too.

Due to the sensitive nature of activation functions, I encourage you all to evaluate its impact on model output before merging. Vectorizing GeLU required trading away a few ulp of worst case accuracy compared to libm. LLMs normally have limitless tolerance for errors, but due to the nature of tanhf() this is a case where even off by ones can cause user-visible changes in model output. It is my belief, based on my own personal experiments so far, that this code works well for llama.cpp, whisper.cpp, gemma, etc.

This software was developed by Mozilla Ocho and ARM Limited. It first appeared in [llamafile](https://github.com/Mozilla-Ocho/llamafile/) which offers you llama.cpp / whisper.cpp / stable-diffusion.cpp with the most bleeding edge performance optimizations and binary distributability.

## Benchmarks

### Intel(R) Core(TM) i9-14900K (AVX2)

Performance is improved by 8x.
Accuracy is improved from 16 to 29 bits.  
Yes GeLU has superior performance on the cheapest microprocessor.

```
master jart@meatball:~/llama.cpp$ make -j tests && ./tests/test-backend-ops -o GELU -b CPU perf

# BEFORE
Backend 1/1 (CPU)
  Backend name: CPU
  GELU(type=f32,ne_a=[128,10,10,10],v=0):   8192 runs -    93.04 us/run -     1000 kB/run -   10.25 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=0):     8192 runs -    20.13 us/run -      310 kB/run -   14.72 GB/s
  GELU(type=f32,ne_a=[128,10,10,10],v=1):   4197 runs -    93.81 us/run -     1999 kB/run -   20.32 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=1):     8191 runs -    20.17 us/run -      621 kB/run -   29.37 GB/s
  Backend CPU: OK

# AFTER
Backend 1/1 (CPU)
  Backend name: CPU
  GELU(type=f32,ne_a=[128,10,10,10],v=0):   8192 runs -    11.82 us/run -     1000 kB/run -   80.66 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=0):     8192 runs -     4.28 us/run -      310 kB/run -   69.28 GB/s
  GELU(type=f32,ne_a=[128,10,10,10],v=1):   4197 runs -    11.65 us/run -     1999 kB/run -  163.68 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=1):     8191 runs -     4.24 us/run -      621 kB/run -  139.82 GB/s
  Backend CPU: OK
```

### Mac Studio M2 Ultra (ARM64)

Performance is improved by 3.1x.
Accuracy is improved from 16 to 29 bits.

```
master jart@studio:~/llama.cpp$ make -j32 tests LLAMA_NO_ACCELERATE=1 LLAMA_NO_METAL=1 && ./tests/test-backend-ops -o GELU -b CPU perf

# BEFORE
Backend 1/1 (CPU)
  Backend name: CPU
  GELU(type=f32,ne_a=[128,10,10,10],v=0):   8192 runs -   125.95 us/run -     1000 kB/run -    7.57 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=0):     8192 runs -    14.57 us/run -      310 kB/run -   20.34 GB/s
  GELU(type=f32,ne_a=[128,10,10,10],v=1):   4197 runs -   120.00 us/run -     1999 kB/run -   15.89 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=1):     8191 runs -    30.37 us/run -      621 kB/run -   19.51 GB/s
  Backend CPU: OK

# AFTER
Backend 1/1 (CPU)
  Backend name: CPU
  GELU(type=f32,ne_a=[128,10,10,10],v=0):   8192 runs -    40.93 us/run -     1000 kB/run -   23.30 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=0):     8192 runs -     9.81 us/run -      310 kB/run -   30.22 GB/s
  GELU(type=f32,ne_a=[128,10,10,10],v=1):   4197 runs -    38.11 us/run -     1999 kB/run -   50.03 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=1):     8191 runs -    10.55 us/run -      621 kB/run -   56.18 GB/s
  Backend CPU: OK
```

### AMD Ryzen Threadripper PRO 7995WX (AVX512)

Performance is improved by 2.4x.
Accuracy is improved from 16 to 29 bits.

```
master jart@luna:~/llama.cpp$ make -j tests && ./tests/test-backend-ops -o GELU -b CPU perf

# BEFORE
Backend 1/1 (CPU)
  Backend name: CPU
  GELU(type=f32,ne_a=[128,10,10,10],v=0):   8192 runs -    27.69 us/run -     1000 kB/run -   34.44 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=0):     8192 runs -    10.24 us/run -      310 kB/run -   28.93 GB/s
  GELU(type=f32,ne_a=[128,10,10,10],v=1):   4197 runs -    27.86 us/run -     1999 kB/run -   68.43 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=1):     8191 runs -    10.20 us/run -      621 kB/run -   58.11 GB/s
  Backend CPU: OK

# AFTER
Backend 1/1 (CPU)
  Backend name: CPU
  GELU(type=f32,ne_a=[128,10,10,10],v=0):   8192 runs -    11.49 us/run -     1000 kB/run -   82.98 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=0):     8192 runs -     7.90 us/run -      310 kB/run -   37.53 GB/s
  GELU(type=f32,ne_a=[128,10,10,10],v=1):   4197 runs -    11.92 us/run -     1999 kB/run -  159.99 GB/s
  GELU(type=f32,ne_a=[7,13,19,23],v=1):     8191 runs -     7.89 us/run -      621 kB/run -   75.07 GB/s
  Backend CPU: OK
```