@jart 2024-08-05
Test failure appears to be unrelated. Something having to do with JSON. `./tests/test-backend-ops -b CPU perf` works fine locally on my Mac Studio.