@JohannesGaessler 2024-08-05
What's the end-to-end speedup?

>Test failure appears to be unrelated. Something having to do with JSON. ./tests/test-backend-ops -b CPU perf works fine locally on my Mac Studio.

According to the log the test failure is caused by `tests/ggml-beckend-ops`. More specifically, a single Metal NMSE value is above the threshold for GELU: `GELU(type=f32,ne_a=[7,13,19,23],v=0): [GELU] NMSE = 0.000001290 > 0.000000100 FAIL`. Notably however the tests use unseeded random inputs so they may fail on some runs but not on other ones. And a threshold of $10^{-7}$ is very strict I think.