@lrvl 2024-08-05
Does llamafile 0.8.12 use this already?