@jart 2024-08-05
Regarding that failure, I've rolled back the bf16 changes to the LUT just to play it safe.

> @lrvl asks: Does llamafile 0.8.12 use this already?

No, you have to build it at HEAD. A release will be coming out shortly. See below for build instructions.

> @JohannesGaessler What's the end-to-end speedup?

For tiny models it can be as high as 17% faster overall. Take for example whisper.cpp using [ggml-tiny-q5_1.bin](https://huggingface.co/ggerganov/whisper.cpp/blob/main/ggml-tiny-q5_1.bin). I can build that with llamafile for the purposes of turning an Edgar Allen Poe wav file into txt as follows:

```
sudo apt install sox libsox-fmt-all
git clone https://github.com/mozilla-ocho/llamafile/
cd llamafile
wget -O whisper-tiny.en-q5_1.bin https://huggingface.co/ggerganov/whisper.cpp/resolve/main/ggml-tiny.en-q5_1.bin
wget https://archive.org/download/raven/raven_poe_64kb.mp3
sox raven_poe_64kb.mp3 -r 16k raven_poe_64kb.wav
make -j o//whisper.cpp
o//whisper.cpp/main -m whisper-tiny.en-q5_1.bin -f raven_poe_64kb.wav
```

And it takes 7.9 seconds. But if I pass the `--trap` flag to enable trapping math, then in llamafile these vectorized libm functions will be disabled.

```
o//whisper.cpp/main -m whisper-tiny.en-q5_1.bin -f raven_poe_64kb.wav --trap
```

And it takes 9.5 seconds. That's for full quality libm tanhf() GeLU. If I use the 16-bit approximation of the GeLU approximation that llama.cpp is currently using, then it does go a little faster: 9.2 seconds. The 128kb LUT actually goes a lot slower in practice than benchmarks would lead you to believe, I think possibly due to how it demolishes the L1 cache.  The vectorized code helps quite a bit with performance on my machines. Plus it gives you near-full accuracy for free.