@ggerganov 2024-08-06
Here are some end-to-end CPU results on my hardware with `llama.cpp` and `whisper.cpp`:

## Gemma-2B

```bash
GGML_NO_METAL=1 ./scripts/compare-commits.sh master pr/8878 \
    -m models/gemma-2b/ggml-model-q4_0.gguf -r 10 -p 0 -t 4,8,12,16
```

| CPU   | Model         |   Threads | Test   |   t/s master |   t/s pr/8878 |   Speedup |
|:------|:--------------|----------:|:-------|-------------:|--------------:|----------:|
| M2 Ultra | gemma 2B Q4_0 |         4 | tg128  |        47.89 |         47.81 |      1.00 |
| M2 Ultra | gemma 2B Q4_0 |         8 | tg128  |        71.46 |         71.09 |      0.99 |
| M2 Ultra | gemma 2B Q4_0 |        12 | tg128  |        83.52 |         82.41 |      0.99 |
| M2 Ultra | gemma 2B Q4_0 |        16 | tg128  |        90.55 |         90.10 |      1.00 |


```bash
./scripts/compare-commits.sh master pr/8878 \
    -m models/gemma-2b/ggml-model-q4_0.gguf -r 10 -p 0 -t 4,8,12,16
```

| CPU                                 | Model         |   Threads | Test   |   t/s master |   t/s pr/8878 |   Speedup |
|:------------------------------------|:--------------|----------:|:-------|-------------:|--------------:|----------:|
| AMD Ryzen 9 5950X 16-Core | gemma 2B Q4_0 |         4 | tg128  |        17.27 |         17.35 |      1.00 |
| AMD Ryzen 9 5950X 16-Core | gemma 2B Q4_0 |         8 | tg128  |        18.28 |         18.50 |      1.01 |
| AMD Ryzen 9 5950X 16-Core | gemma 2B Q4_0 |        12 | tg128  |        18.38 |         18.54 |      1.01 |
| AMD Ryzen 9 5950X 16-Core | gemma 2B Q4_0 |        16 | tg128  |        18.33 |         18.40 |      1.00 |



## Whisper

```bash
GGML_NO_METAL=1 make -j && ./scripts/bench-all.sh 8
```

- `master`

|    CPU |     Config |         Model |  Th |  FA |    Enc. |    Dec. |    Bch5 |      PP |  Commit |
|    --- |        --- |           --- | --- | --- |     --- |     --- |     --- |     --- |     --- |
| M2 Ultra |  NEON BLAS |          tiny |   8 |   0 |   80.80 |    1.06 |    0.35 |    0.11 | fe36c90 |
| M2 Ultra |  NEON BLAS |     tiny-q5_0 |   8 |   0 |   80.81 |    0.76 |    0.34 |    0.11 | fe36c90 |
| M2 Ultra |  NEON BLAS |     tiny-q5_1 |   8 |   0 |   83.11 |    0.80 |    0.38 |    0.11 | fe36c90 |
| M2 Ultra |  NEON BLAS |          base |   8 |   0 |  159.92 |    1.83 |    0.65 |    0.21 | fe36c90 |
| M2 Ultra |  NEON BLAS |     base-q5_0 |   8 |   0 |  160.57 |    1.32 |    0.63 |    0.21 | fe36c90 |
| M2 Ultra |  NEON BLAS |     base-q5_1 |   8 |   0 |  150.90 |    1.37 |    0.67 |    0.21 | fe36c90 |
| M2 Ultra |  NEON BLAS |         small |   8 |   0 |  466.81 |    4.69 |    1.73 |    0.56 | fe36c90 |
| M2 Ultra |  NEON BLAS |    small-q5_0 |   8 |   0 |  462.21 |    3.29 |    1.63 |    0.55 | fe36c90 |
| M2 Ultra |  NEON BLAS |    small-q5_1 |   8 |   0 |  469.31 |    3.41 |    1.77 |    0.56 | fe36c90 |


- cherry-pick this PR in `whisper.cpp`

|    CPU |     Config |         Model |  Th |  FA |    Enc. |    Dec. |    Bch5 |      PP |  Commit |
|    --- |        --- |           --- | --- | --- |     --- |     --- |     --- |     --- |     --- |
| M2 Ultra |  NEON BLAS |          tiny |   8 |   0 |   80.47 |    1.07 |    0.35 |    0.11 | fe36c90 |
| M2 Ultra |  NEON BLAS |     tiny-q5_0 |   8 |   0 |   79.52 |    0.77 |    0.35 |    0.11 | fe36c90 |
| M2 Ultra |  NEON BLAS |     tiny-q5_1 |   8 |   0 |   76.88 |    0.76 |    0.36 |    0.11 | fe36c90 |
| M2 Ultra |  NEON BLAS |          base |   8 |   0 |  157.43 |    1.81 |    0.64 |    0.21 | fe36c90 |
| M2 Ultra |  NEON BLAS |     base-q5_0 |   8 |   0 |  157.88 |    1.29 |    0.62 |    0.21 | fe36c90 |
| M2 Ultra |  NEON BLAS |     base-q5_1 |   8 |   0 |  154.50 |    1.35 |    0.67 |    0.21 | fe36c90 |
| M2 Ultra |  NEON BLAS |         small |   8 |   0 |  461.85 |    4.63 |    1.72 |    0.57 | fe36c90 |
| M2 Ultra |  NEON BLAS |    small-q5_0 |   8 |   0 |  450.27 |    3.25 |    1.62 |    0.56 | fe36c90 |
| M2 Ultra |  NEON BLAS |    small-q5_1 |   8 |   0 |  454.20 |    3.46 |    1.76 |    0.56 | fe36c90 |

*(These are times in `ms` (lower is better). Would have expected to see a slight gain in `Dec` and `Bch5`)*

Overall, I don't see much difference in the end-to-end performance. I do observe the speed improvement of the individual GELU op measured by `test-backend-ops perf`, but it seems it's quite negligible to the overall model evaluation. @jart can you run Gemma-2B comparison on your hardware and report the numbers?

> For tiny models it can be as high as 17% faster overall.

For measuring the performance of `whisper.cpp` I recommend to use the `bench` tool, or the `scripts/bench-all.sh` script as I did above. Measuring the total transcription time of auto-regressive models is meaningless. Apart from `whisper.cpp` I have never seen anyone else report the performance results correctly for Whisper